move constant values to config folder
