+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export const ANALYTICS_ENDPOINT = 'https://api.hackertab.dev/analytics'
+export const ANALYTICS_SDK_KEY = '9662c93f91473ba6e96711b22e0a367d'
+export const LS_ANALYTICS_ID_KEY = 'hackerTabAnalyticsId'
